Use data/types.h

+++ user/include/sys/calls.h
@@ -1,11 +1,12 @@
 #ifndef USER_SYSTEM_CALLS_H
 #define USER_SYSTEM_CALLS_H
 
+#include <data/types.h>
+
 void exit();
 char getchar();
 void putchar(unsigned char c);
-void wait(unsigned long duration);
-void create_thread(void (*func)(void*), const void* args,
-                   unsigned long args_size);
+void wait(uint32 duration);
+void create_thread(void (*func)(void*), const void* args, uint32 args_size);
 
 #endif /* USER_SYSTEM_CALLS_H */
